[native] Use NavigationHandler and RootRouter to handle log outs
