+++ native/navigation/root-router.js
@@ -10,7 +10,12 @@ import {
 } from 'react-navigation';
 
 import { removeScreensFromStack } from '../utils/navigation-utils';
-import { accountModals } from './route-names';
+import {
+  accountModals,
+  LoggedOutModalRouteName,
+  AppRouteName,
+} from './route-names';
+import { defaultNavInfo } from './default-state';
 
 type LogInAction = {|
   +type: 'LOG_IN',
@@ -52,7 +57,54 @@ function RootRouter(
           isTransitioning: true,
         };
       } else if (action.type === 'LOG_OUT') {
-        return lastState;
+        if (!lastState) {
+          return lastState;
+        }
+        let newState = { ...lastState };
+        newState.routes[0] = defaultNavInfo.navigationState.routes[0];
+
+        const initialKey = newState.routes[newState.index].key;
+        let loggedOutModalFound = false;
+        newState = removeScreensFromStack(
+          newState,
+          (route: NavigationRoute) => {
+            const { routeName } = route;
+            if (routeName === LoggedOutModalRouteName) {
+              loggedOutModalFound = true;
+            }
+            return routeName === AppRouteName ||
+              accountModals.includes(routeName)
+              ? 'keep'
+              : 'remove';
+          },
+        );
+
+        let isTransitioning =
+          newState.routes[newState.index].key === initialKey;
+        if (!loggedOutModalFound) {
+          const [appRoute, ...restRoutes] = newState.routes;
+          newState = {
+            ...newState,
+            index: newState.index + 1,
+            routes: [
+              appRoute,
+              { key: 'LoggedOutModal', routeName: LoggedOutModalRouteName },
+              ...restRoutes,
+            ],
+          };
+          if (newState.index === 1) {
+            isTransitioning = true;
+          }
+        }
+
+        if (isTransitioning) {
+          newState = {
+            ...newState,
+            isTransitioning,
+          };
+        }
+
+        return newState;
       } else {
         return stackRouter.getStateForAction(action, lastState);
       }
